Added some javadocs
