--- simulator-core/src/main/java/com/evbox/everon/ocpp/simulator/station/component/SetVariableHandler.java
@@ -1,10 +1,0 @@
-package com.evbox.everon.ocpp.simulator.station.component;
-
-import com.evbox.everon.ocpp.v20.message.centralserver.SetVariableDatum;
-
-@FunctionalInterface
-public interface SetVariableHandler {
-
-    void handle(SetVariableDatum setVariableDatum);
-
-}
