+++ simulator-core/src/main/java/com/evbox/everon/ocpp/simulator/station/component/exception/EVSEComponent.java
@@ -4,6 +4,9 @@ import com.evbox.everon.ocpp.simulator.station.Station;
 import com.evbox.everon.ocpp.simulator.station.component.StationComponent;
 import com.google.common.collect.ImmutableList;
 
+/**
+ * Representation of EVSE component according to OCPP 2.0 (3.2.30. EVSE)
+ */
 public class EVSEComponent extends StationComponent {
 
     public static final String NAME = "EVSE";
