+++ simulator-core/src/main/java/com/evbox/everon/ocpp/simulator/station/component/ocppcommctrlr/OCPPCommCtrlrComponent.java
@@ -4,6 +4,9 @@ import com.evbox.everon.ocpp.simulator.station.Station;
 import com.evbox.everon.ocpp.simulator.station.component.StationComponent;
 import com.google.common.collect.ImmutableList;
 
+/**
+ * Representation of OCPPCommCtrlr component according to OCPP 2.0 (3.1.7. OCPPCommCtrlr)
+ */
 public class OCPPCommCtrlrComponent extends StationComponent {
 
     public static final String NAME = "OCPPCommCtrlr";
